DTMT-19 - create driver-mongo module and move related src
