+++ modules/core/src/main/kotlin/datamaintain/core/step/sort/ByLengthAndCaseInsensitiveAlphabeticalSorter.kt
@@ -1,8 +1,7 @@
-package datamaintain.sort
+package datamaintain.core.step.sort
 
 import datamaintain.core.Config
 import datamaintain.core.script.Script
-import datamaintain.core.step.sort.Sorter
 
 class ByLengthAndCaseInsensitiveAlphabeticalSorter(config: Config) : Sorter<String>(config) {
     override fun <T : Script> sort(scripts: List<T>, getter: (T) -> String): List<T> {
