--- modules/core/src/test/resources/executor_test_files/mongo/mongo_simple_insert.js
@@ -1,1 +1,0 @@
-db.simple.insert({ find: "me", data: 'inserted'});
