+++ modules/driver-mongo/src/test/kotlin/datamaintain/db/driver/mongo/test/AbstractMongoDbTest.kt
@@ -1,4 +1,4 @@
-package datamaintain.test
+package datamaintain.db.driver.mongo.test
 
 import datamaintain.core.script.ScriptWithoutContent
 import datamaintain.db.driver.mongo.MongoDatamaintainDriver
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.litote.kmongo.KMongo
 
-abstract class AbstractDbTest {
+abstract class AbstractMongoDbTest {
     val databaseName = "datamaintain-test"
     val mongoUri = "localhost:27018"
